+++ src/shared/git-rebase-source.ts
@@ -1,3 +1,6 @@
+// Why: a stalled remote must fail the rebase fetch, not hang the rebase; client and relay share one bound.
+export const REBASE_SOURCE_FETCH_TIMEOUT_MS = 60_000
+
 export type GitCommandRunner = (args: string[]) => Promise<{ stdout: string }>
 
 export type GitRemoteRebaseSource = {
